restructured and fixed errors, created strange color issue
