+++ ent.h
